transformer/app.py
import json
import os
import re
from dataclasses import asdict, dataclass
from typing import Optional

import chromadb
import requests
from dotenv import load_dotenv
from flask import Flask, jsonify, request, send_from_directory
from flask_cors import CORS

from sentence_transformers import SentenceTransformer

load_dotenv()

EMBEDDING_MODEL = "all-MiniLM-L6-v2"
COLLECTION_NAME = "news"
COSINE_THRESHOLD = 0.57
DEFAULT_MODEL = "mistral-small-latest"
DEFAULT_API_URL = "https://api.mistral.ai/v1/chat/completions"
DEFAULT_TRUNCATION = 1000

MISTRAL_API_KEY = os.getenv("MISTRAL_API_KEY")
MISTRAL_MODEL = os.getenv("MISTRAL_MODEL", DEFAULT_MODEL)
MISTRAL_API_URL = os.getenv("MISTRAL_API_URL", DEFAULT_API_URL)

# Load embedding model
embedding_model = SentenceTransformer(EMBEDDING_MODEL)

# Set up ChromaDB with correct path
BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
CHROMADB_PATH = os.path.join(BASE_DIR, "data", "chroma_db")
os.makedirs(CHROMADB_PATH, exist_ok=True)

client = chromadb.PersistentClient(path=CHROMADB_PATH)
collection = client.get_or_create_collection(
    name=COLLECTION_NAME,
    metadata={"hnsw:space": "cosine"},
)

FRONTEND_DIR = os.path.join(BASE_DIR, "frontend")
app = Flask(__name__, static_folder=FRONTEND_DIR, template_folder=FRONTEND_DIR)
CORS(app)

PROMPT_TEMPLATE = """
You are a misinformation detection expert.

You are given news snippets retrieved from a vector database based on the input text provided.
Use ONLY these snippets as evidence when judging the credibility of that text.

Snippets:
\"\"\"{content}\"\"\"

Step 1: Identify if the information in these snippets appears misleading, exaggerated, or false.
Step 2: Detect patterns typical of misinformation.
Step 3: Output a credibility score from 0 (false) to 100 (credible) and a reason.

Respond in this format only:
Credibility Score: <score>
Reason: <brief reason>
""".strip()

_RE_SCORE  = re.compile(r"Credibility Score:\s*\**(\d+)\**", re.IGNORECASE)
_RE_REASON = re.compile(r"Reason:\s*(.*)", re.IGNORECASE | re.DOTALL)
URL_PATTERN = re.compile(r"https?://[^\s\"')>\]]+")



@dataclass
class CredibilityResult:
    article: str
    snippets: str
    score: Optional[int]
    reason: str

    def serialize(self, truncate_at: int = DEFAULT_TRUNCATION) -> dict:
        payload = asdict(self)
        if self.article:
            payload["article"] = self.article[: truncate_at or DEFAULT_TRUNCATION]
        if self.snippets:
            payload["snippets"] = self.snippets[: truncate_at or DEFAULT_TRUNCATION]
        payload["score"] = self.score
        return payload


def extract_article_from_text(text: str) -> str:
    """Validate and clean input text."""
    if not text or not text.strip():
        raise ValueError("Article text cannot be empty.")
    return text.strip()


def _get_embedding(text: str) -> list[float]:
    """Generate embedding using sentence-transformers."""
    embedding = embedding_model.encode(text)
    return embedding.tolist()


def retrieve(query: str, top_n: int = 10) -> tuple[list[str], list[str]]:
    """Embed the query and retrieve similar docs from ChromaDB."""
    qe = _get_embedding(query)
    results = collection.query(
        query_embeddings=[qe], n_results=top_n, include=["documents", "distances"]
    )

    docs = results.get("documents", [[]])[0]
    distances = results.get("distances", [[]])[0]

    filtered_docs = []
    filtered_links = []
    
    for doc, distance in zip(docs, distances):
        if distance is not None and (1 - distance) >= COSINE_THRESHOLD:
            filtered_docs.append(doc)
            link = extract_reference_link(doc)
            if link:
                filtered_links.append(link)

    return filtered_docs, filtered_links


def extract_reference_link(doc):
    if not doc:
        return None

    if isinstance(doc, dict):
        for key in ("link", "url", "source"):
            value = doc.get(key)
            if isinstance(value, str) and value.startswith("http"):
                return value

    if isinstance(doc, str):
        try:
            parsed = json.loads(doc)
            if isinstance(parsed, dict):
                ref = parsed.get("link") or parsed.get("url")
                if isinstance(ref, str) and ref.startswith("http"):
                    return ref
        except json.JSONDecodeError:
            pass

        match = URL_PATTERN.search(doc)
        if match:
            return match.group(0).rstrip(".,);")

    return None


def _parse_model_response(content: str) -> tuple[Optional[int], str]:
    if not content:
        return None, "Empty response from model."

    score_match  = _RE_SCORE.search(content)
    reason_match = _RE_REASON.search(content)

    if not score_match:
        print(f"[PHNX] Could not parse score from model response:\n{content}\n---")
        return None, "Unable to parse credibility score from model response."

    score  = int(score_match.group(1))
    reason = reason_match.group(1).strip() if reason_match else "No reason provided."
    return score, reason


def check_credibility(snippets: str) -> CredibilityResult:
    if not MISTRAL_API_KEY:
        raise RuntimeError(
            "MISTRAL_API_KEY is not set. Please define it in your environment or .env file."
        )

    payload = {
        "model": MISTRAL_MODEL,
        "messages": [{"role": "user", "content": PROMPT_TEMPLATE.format(content=snippets)}],
        "temperature": 0,
    }

    response = requests.post(
        MISTRAL_API_URL,
        headers={
            "Authorization": f"Bearer {MISTRAL_API_KEY}",
            "Content-Type": "application/json",
        },
        json=payload,
        timeout=60,
    )
    if response.status_code != 200:
        raise RuntimeError(
            f"Mistral API error {response.status_code}: {response.text[:200]}"
        )

    data = response.json()
    content = data.get("choices", [{}])[0].get("message", {}).get("content", "")
    score, reason = _parse_model_response(content)
    # article will be filled in api_check, we just keep the structure here
    return CredibilityResult(article="", snippets=snippets, score=score, reason=reason)


@app.route("/")
def index():
    return send_from_directory(FRONTEND_DIR, "index.html")


@app.post("/api/check")
def api_check():
    data = request.get_json(silent=True) or {}
    text = (data.get("text") or "").strip()
    top_n = data.get("top_n", 3)

    if not text:
        return jsonify({"error": "Article text is required."}), 400

    try:
        # Clean input text
        article = extract_article_from_text(text)

        try:
            top_n = max(1, int(top_n))
        except (TypeError, ValueError):
            top_n = 3

        # Use the article text as the query into your vector DB
        documents, reference_links = retrieve(article, top_n=top_n)
        snippets = " ".join(documents) if documents else ""
        result = check_credibility(snippets or article)

        payload = result.serialize()
        payload["article"] = article[:DEFAULT_TRUNCATION]
        payload["documents"] = reference_links
        return jsonify(payload)
    except Exception as exc:
        return jsonify({"error": str(exc)}), 500


#import os

if __name__ == "__main__":
    port = int(os.environ.get("TRANSFORMER_PORT", 5001))
    app.run(host="0.0.0.0", port=port)
